moving to scenario utils
